moved common functions to separate file
