+++ src/components/post-card/PostCard.jsx
@@ -17,6 +17,13 @@ import {
 } from "@heroicons/react/outline";
 import { EditPostModal } from "../modal/EditPostModal";
 import { likePost, dislikePost, deletePostThunk } from "../../redux/postSlice";
+import {
+  checkBookmarkedPost,
+  checkLikedPost,
+  checkPostIdOfLoggedUser,
+  checkUserIsFollowed,
+  getTimeInDMY,
+} from "../../utility/commonFunctions";
 
 const PostCard = ({ post, cardType }) => {
   const {
@@ -36,13 +43,6 @@ const PostCard = ({ post, cardType }) => {
   const { user } = useSelector((state) => state.authSlice);
   const [profileImageLoaded, setProfileImageLoaded] = useState(false);
   const dispatch = useDispatch();
-  const checkUserIsFollowed = () => {
-    return (
-      user.following.find(
-        (userFollowing) => userFollowing.username === username
-      ) || username === user.username
-    );
-  };
 
   const followUser = async () => {
     dispatch(followUnFollowUserUserThunk({ userId, type: "follow" }));
@@ -70,31 +70,7 @@ const PostCard = ({ post, cardType }) => {
   const editPost = () => {
     setShowEditPostModal(true);
   };
-  const checkLikedPost = () => {
-    const likedArray = likes.likedBy.filter(
-      (userLiked) => userLiked._id === user._id
-    );
-    if (likedArray.length === 0) return false;
-    else return true;
-  };
-  const checkBookmarkedPost = () => {
-    const bookmaredArray = user.bookmarks.filter((post) => post._id === _id);
-    if (bookmaredArray.length === 0) return false;
-    else return true;
-  };
-  const checkPostIdOfLoggedUser = () => {
-    return username === user.username;
-  };
-  const getTimeInDMY = (date) => {
-    const dateF = new Date(date);
-    const dmy =
-      dateF.getDate() +
-      " " +
-      dateF.toLocaleString("default", { month: "long" }) +
-      ", " +
-      dateF.getFullYear();
-    return dmy;
-  };
+
   return (
     <div className=" p-10 flex flex-col gap-5 border-t border-b bg-primary-bg-color w-full">
       {showEditPostModal && (
@@ -132,7 +108,7 @@ const PostCard = ({ post, cardType }) => {
             <p className="text-xs text-gray-600">{getTimeInDMY(createdAt)}</p>
           </div>
         </Link>
-        {checkPostIdOfLoggedUser() && (
+        {checkPostIdOfLoggedUser({ user, username }) && (
           <div className="flex flex-row">
             <button onClick={deletePost}>
               <TrashIcon className="p-2 h-10 w-10 hover:bg-hover-color rounded-full" />
@@ -144,7 +120,7 @@ const PostCard = ({ post, cardType }) => {
         )}
 
         {!(username === user.username) &&
-          (checkUserIsFollowed() ? (
+          (checkUserIsFollowed({ user, username }) ? (
             <button
               onClick={unfollowUser}
               className="border p-1 px-2 rounded-md hover:bg-hover-color"
@@ -180,7 +156,7 @@ const PostCard = ({ post, cardType }) => {
         )}
       </div>
       <div className="flex flex-row gap-20">
-        {checkLikedPost() ? (
+        {checkLikedPost({ likes, user }) ? (
           <div
             className="flex flex-row items-center cursor-pointer gap-2"
             onClick={dislikePostFunction}
@@ -197,7 +173,7 @@ const PostCard = ({ post, cardType }) => {
             <p>{likes.likedBy.length}</p>
           </div>
         )}
-        {checkBookmarkedPost() ? (
+        {checkBookmarkedPost({ user, _id }) ? (
           <div
             className="flex flex-row items-center cursor-pointer gap-2"
             onClick={removeBookmarkPostFunction}
